mobile reader: extract useReaderProgress hook
Owns saveProgress callback + unmount/AppState flush effects.
Same offline-first behavior: local persist always, server PUT for
authenticated users only.
